[ccore] Refactoring of agglomerative algorithm and its interface to pyclustering.

+++ pyclustering/core/agglomerative_wrapper.py
@@ -23,7 +23,7 @@
 
 """
 
-from ctypes import cdll, c_uint;
+from ctypes import cdll, c_size_t;
 
 from pyclustering.core.wrapper import PATH_DLL_CCORE_64, create_pointer_data, extract_pyclustering_package;
 
@@ -31,7 +31,7 @@ def agglomerative_algorithm(data, number_clusters, link):
     pointer_data = create_pointer_data(data);
 
     ccore = cdll.LoadLibrary(PATH_DLL_CCORE_64);
-    package = ccore.agglomerative_algorithm(pointer_data, c_uint(number_clusters), c_uint(link));
+    package = ccore.agglomerative_algorithm(pointer_data, c_size_t(number_clusters), c_size_t(link));
 
     result = extract_pyclustering_package(package);
     ccore.free_pyclustering_package(package);
